Rename & move UI token classes
Shared with API
